Refactoring models and association table
